[MWB] - moving MyRectangle from Common.MachineStuff.cs into MyRectangle.cs - #35155
